Consolidate tag components
